requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
fastapi==0.103.1
uvicorn==0.23.2
mlflow==2.7.1
pytest==7.4.2
pyyaml==6.0.1
httpx==0.24.1
python-dotenv==1.0.0
joblib==1.3.2
email-validator>=2.0.0